requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
billiard==3.3.0.23
celery==3.1.26.post2
Django==3.2.9
django-celery==3.3.1
django-environ==0.8.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
kombu==3.0.37
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
